requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.3.0=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
babel=2.15.0=pypi_0
backoff=2.2.1=pypi_0
beautifulsoup4=4.12.3=pypi_0
blis=0.7.11=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.3.11=haa95532_0
catalogue=2.0.10=pypi_0
certifi=2024.2.2=pypi_0
cffi=1.16.0=pypi_0
chardet=5.2.0=pypi_0
charset-normalizer=3.3.2=pypi_0
chromedriver-autoinstaller=0.6.4=pypi_0
click=8.1.7=pypi_0
cloudpathlib=0.16.0=pypi_0
colorama=0.4.6=pypi_0
coloredlogs=15.0.1=pypi_0
confection=0.1.4=pypi_0
courlan=1.1.0=pypi_0
cssselect=1.2.0=pypi_0
cymem=2.0.8=pypi_0
dataclasses-json=0.6.6=pypi_0
dateparser=1.2.0=pypi_0
deepdiff=7.0.1=pypi_0
deprecated=1.2.14=pypi_0
dirtyjson=1.0.8=pypi_0
distro=1.9.0=pypi_0
dnspython=2.6.1=pypi_0
email-validator=2.1.1=pypi_0
emoji=2.12.1=pypi_0
exceptiongroup=1.2.1=pypi_0
fastapi=0.111.0=pypi_0
fastapi-cli=0.0.4=pypi_0
fastembed=0.2.7=pypi_0
feedfinder2=0.0.4=pypi_0
feedparser=6.0.11=pypi_0
filelock=3.14.0=pypi_0
filetype=1.2.0=pypi_0
flatbuffers=24.3.25=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.5.0=pypi_0
greenlet=3.0.3=pypi_0
groq=0.7.0=pypi_0
grpcio=1.64.0=pypi_0
grpcio-tools=1.64.0=pypi_0
h11=0.14.0=pypi_0
h2=4.1.0=pypi_0
hpack=4.0.0=pypi_0
html2text=2020.1.16=pypi_0
htmldate=1.8.1=pypi_0
httpcore=1.0.5=pypi_0
httptools=0.6.1=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.23.1=pypi_0
humanfriendly=10.0=pypi_0
hyperframe=6.0.1=pypi_0
idna=3.7=pypi_0
intel-openmp=2021.4.0=pypi_0
jieba3k=0.35.1=pypi_0
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
jsonpatch=1.33=pypi_0
jsonpath-ng=1.6.1=pypi_0
jsonpath-python=1.0.6=pypi_0
jsonpointer=2.4=pypi_0
justext=3.0.1=pypi_0
langchain=0.2.0=pypi_0
langchain-community=0.2.0=pypi_0
langchain-core=0.2.0=pypi_0
langchain-text-splitters=0.2.0=pypi_0
langcodes=3.4.0=pypi_0
langdetect=1.0.9=pypi_0
langsmith=0.1.60=pypi_0
language-data=1.2.0=pypi_0
libffi=3.4.4=hd77b12b_1
llama-index=0.10.37=pypi_0
llama-index-agent-openai=0.2.5=pypi_0
llama-index-cli=0.1.12=pypi_0
llama-index-core=0.10.37.post1=pypi_0
llama-index-embeddings-ollama=0.1.2=pypi_0
llama-index-embeddings-openai=0.1.10=pypi_0
llama-index-indices-managed-llama-cloud=0.1.6=pypi_0
llama-index-legacy=0.9.48=pypi_0
llama-index-llms-groq=0.1.4=pypi_0
llama-index-llms-openai=0.1.20=pypi_0
llama-index-llms-openai-like=0.1.3=pypi_0
llama-index-multi-modal-llms-openai=0.1.6=pypi_0
llama-index-postprocessor-colbert-rerank=0.1.2=pypi_0
llama-index-program-openai=0.1.6=pypi_0
llama-index-question-gen-openai=0.1.3=pypi_0
llama-index-readers-file=0.1.22=pypi_0
llama-index-readers-llama-parse=0.1.4=pypi_0
llama-index-readers-web=0.1.16=pypi_0
llama-index-readers-wikipedia=0.1.4=pypi_0
llama-index-vector-stores-qdrant=0.2.8=pypi_0
llama-parse=0.4.3=pypi_0
llamaindex-py-client=0.1.19=pypi_0
loguru=0.7.2=pypi_0
lxml=5.1.1=pypi_0
maincontentextractor=0.0.4=pypi_0
marisa-trie=1.1.1=pypi_0
markdown=3.6=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
marshmallow=3.21.2=pypi_0
mdurl=0.1.2=pypi_0
mkl=2021.4.0=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
murmurhash=1.0.10=pypi_0
mypy-extensions=1.0.0=pypi_0
nest-asyncio=1.6.0=pypi_0
networkx=3.3=pypi_0
newspaper3k=0.2.8=pypi_0
nltk=3.8.1=pypi_0
numpy=1.26.4=pypi_0
ollama=0.2.0=pypi_0
onnx=1.16.0=pypi_0
onnxruntime=1.18.0=pypi_0
openai=1.30.1=pypi_0
openssl=1.1.1w=h2bbff1b_0
ordered-set=4.1.0=pypi_0
orjson=3.10.3=pypi_0
outcome=1.3.0.post0=pypi_0
packaging=23.2=pypi_0
pandas=2.2.2=pypi_0
pillow=10.3.0=pypi_0
pip=24.0=py310haa95532_0
playwright=1.44.0=pypi_0
ply=3.11=pypi_0
portalocker=2.8.2=pypi_0
preshed=3.0.9=pypi_0
protobuf=5.26.1=pypi_0
pycparser=2.22=pypi_0
pydantic=2.7.1=pypi_0
pydantic-core=2.18.2=pypi_0
pyee=11.1.0=pypi_0
pygments=2.18.0=pypi_0
pypdf=4.2.0=pypi_0
pyreadline3=3.4.1=pypi_0
pysocks=1.7.1=pypi_0
python=3.10.0=h96c0403_3
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python-iso639=2024.4.27=pypi_0
python-magic=0.4.27=pypi_0
python-multipart=0.0.9=pypi_0
pytz=2024.1=pypi_0
pywin32=306=pypi_0
pyyaml=6.0.1=pypi_0
qdrant-client=1.9.1=pypi_0
rapidfuzz=3.9.1=pypi_0
regex=2024.5.15=pypi_0
requests=2.32.1=pypi_0
requests-file=2.0.0=pypi_0
rich=13.7.1=pypi_0
safetensors=0.4.3=pypi_0
scikit-learn=1.5.0=pypi_0
scipy=1.13.0=pypi_0
selenium=4.21.0=pypi_0
sentence-transformers=2.7.0=pypi_0
setuptools=69.5.1=py310haa95532_0
sgmllib3k=1.0.0=pypi_0
shellingham=1.5.4=pypi_0
six=1.16.0=pypi_0
smart-open=6.4.0=pypi_0
sniffio=1.3.1=pypi_0
sortedcontainers=2.4.0=pypi_0
soupsieve=2.5=pypi_0
spacy=3.7.4=pypi_0
spacy-legacy=3.0.12=pypi_0
spacy-loggers=1.0.5=pypi_0
spider-client=0.0.11=pypi_0
sqlalchemy=2.0.30=pypi_0
sqlite=3.45.3=h2bbff1b_0
srsly=2.4.8=pypi_0
starlette=0.37.2=pypi_0
striprtf=0.0.26=pypi_0
sympy=1.12=pypi_0
tabulate=0.9.0=pypi_0
tbb=2021.12.0=pypi_0
tenacity=8.3.0=pypi_0
thinc=8.2.3=pypi_0
threadpoolctl=3.5.0=pypi_0
tiktoken=0.7.0=pypi_0
tinysegmenter=0.3=pypi_0
tk=8.6.14=h0416ee5_0
tld=0.13=pypi_0
tldextract=5.1.2=pypi_0
tokenizers=0.19.1=pypi_0
torch=2.3.0=pypi_0
tqdm=4.66.4=pypi_0
trafilatura=1.9.0=pypi_0
transformers=4.41.0=pypi_0
trio=0.25.1=pypi_0
trio-websocket=0.11.1=pypi_0
typer=0.12.3=pypi_0
typing-extensions=4.11.0=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2024.1=pypi_0
tzlocal=5.2=pypi_0
ujson=5.10.0=pypi_0
unstructured=0.14.0=pypi_0
unstructured-client=0.22.0=pypi_0
urllib3=2.2.1=pypi_0
uvicorn=0.29.0=pypi_0
vc=14.2=h2eaa2aa_1
vs2015_runtime=14.29.30133=h43f2093_3
wasabi=1.1.2=pypi_0
watchfiles=0.21.0=pypi_0
weasel=0.3.4=pypi_0
websockets=12.0=pypi_0
wheel=0.43.0=py310haa95532_0
wikipedia=1.4.0=pypi_0
win32-setctime=1.1.0=pypi_0
wrapt=1.16.0=pypi_0
wsproto=1.2.0=pypi_0
xz=5.4.6=h8cc25b3_1
yarl=1.9.4=pypi_0
zlib=1.2.13=h8cc25b3_1
